minor fix QBItemClickListener
